+++ pptx/slide.go
@@ -482,6 +482,10 @@ func (s *Slide) syncShapesToXML() {
 			id := allocID()
 			spTree.GraphicFrame = append(spTree.GraphicFrame, smartArtFrameToOxml(sh, id))
 			sh.sourceID = id
+		case *OLEObjectFrame:
+			id := allocID()
+			spTree.GraphicFrame = append(spTree.GraphicFrame, oleObjectFrameToOxml(sh, id))
+			sh.sourceID = id
 		case *Picture:
 			id := allocID()
 			spTree.Pic = append(spTree.Pic, pictureToOxml(sh, id))
@@ -632,6 +636,9 @@ func (s *Slide) appendShapesToXML(spTree *oxml.ShapeTree, shapes []Shape) {
 			ref = oxml.ChildRef{Kind: oxml.ChildGraphicFrame, Index: len(spTree.GraphicFrame) - 1}
 		case *SmartArtFrame:
 			spTree.AppendGraphicFrame(smartArtFrameToOxml(sh, id))
+			ref = oxml.ChildRef{Kind: oxml.ChildGraphicFrame, Index: len(spTree.GraphicFrame) - 1}
+		case *OLEObjectFrame:
+			spTree.AppendGraphicFrame(oleObjectFrameToOxml(sh, id))
 			ref = oxml.ChildRef{Kind: oxml.ChildGraphicFrame, Index: len(spTree.GraphicFrame) - 1}
 		case *Picture:
 			spTree.AppendPic(pictureToOxml(sh, id))